refactor(arch): break symbol_query ↔ symbols::handlers cycle (PR-F)
review_architecture (self-introspect) reported a 3-node cycle:
mod.rs → ranked_context.rs → symbols/handlers.rs → (back via the
pipeline tool entries). The upward edge came from the pipeline
importing `sparse_symbol_hits_for_query` + `adapt_budget_to_context_window`
from `handlers.rs`.

Move both helpers (plus the `flatten_symbols` utility they depend on,
and the `adapt_budget_tests` block) into a new
`tools/symbol_query/sparse_retriever.rs`. After this PR the dependency
flow is one-way: `symbols::handlers` → `symbol_query::sparse_retriever`.
The pipeline owns its retrieval primitives end-to-end; the legacy
BM25 / get_complexity handlers in `handlers.rs` reach down into the
pipeline module for the shared bits.

handlers.rs -130 LOC. CLAUDE.md "Symbol-query path lives behind one
seam" updated: the previous "intentional, not a tightening target"
note for sparse helpers is replaced with the new sparse_retriever
seam description + an explicit "never reach upward from the pipeline
into symbols::handlers" rule.

Validation:
- cargo clippy --workspace --features http,semantic -- -D warnings: clean
- cargo clippy --workspace --no-default-features -- -D warnings: clean
- cargo clippy --workspace --features scip-backend -- -D warnings: clean
- cargo test -p codelens-mcp --bin codelens-mcp: 593 passed, 0 failed
- adapt_budget_tests (6 tests) pass at new location

Live-daemon cycle check will run after the binary rebuild that follows
the merge.
